backend/requirements.txt
flask
flask-cors
gunicorn
aiohttp
apscheduler
coinbase==2.1.0
flasgger
flask_sqlalchemy
python-dotenv
